Enable referenecs to refer to labels
References now support using a label as a base point of reference. Label references may be offset similar to instruction pointer references.

The `ALTER` command is likewise updated to be able to add or remove labels. The command now adds a label if one does not already exist and removes a label from a command if the label name is `UNDEFINED` rather than a string.