Refactor Out Program State
This change updates the internals of handling program state by codifying that state into an well-defined object. Instructions now behave as transformations against that state.

The largest change that came as a result of this refactoring is how references are processed. The logic has been split into two parts: dereferencing a single reference and the effects that that operation has on the state and aggregating that processing for instructions which take multiple references.

To better handle references and instruction data, instruction data is no longer always an array. Instruction data now is a STOP type. This effects the `PUSH` and `INJECT` commands as they now are implemented by generating a command string given their input data and then parsing that command as all other commands are parsed.

As part of this work a bug in the way absolute references were converted to strings was fixed.