Make the interpreter a resumable state machine
This change lays the foundation for the interpreter to be stopped and started in the middle of the program. The interpreter's state now carries all information so that it may be stopped whenever a new instruction is executed.

This change had the largest effect on how references are handled. Indirect references and references to the current value of the instruction pointer may be evaluated in the same step as evaluating the instruction itself but evaluating direct references causes new instructions to be executed. A stack of instructions and the state of their references and their location in the set of instructions must be maintained separate from the set of instructions.

Each time an instruction is popped from this stack the return value populates the next direct reference. When the last instruction is popped from this stack the instruction at the instruction pointer is automatically pushed on to the stack.