Support referencing the current command
This change differentiates between the location of the instruction pointer and the location of the current command. The instruction pointer can only be modified by root-level commands and so always references the location of the currently executing root level command. In contrast, the location of the current command is also an interesting location to know about.

These two concepts are represented by the `$ip` and `$ci` references, respectively.