Allow programs to be paused and reset
This change further breaks up the way that STOP programs have so far been run. The `execute` method is not flexible enough to allow inspection into a running program because it offers no chance for intermediate state to be observed.

This change replaces the `execute` method with a new callback which accepts the new program state and returns whether the program should pause and with a new method `go` which advances the program's state until instructed to pause by the callback.

Since resetting the program is an interesting operation once you can begin to inspect its state a new `reset` method is added to support this as well.