Replace `READ` with `$stdin`
Since the `READ` instruction returns a value but never takes in any arguments it never made sense to have more than one `READ` instruction per program. Typical usage would be to write

`(READ) READ`

and then to refer to it using `$READ`. This change codifies that usage into a special reference, `$stdin`, which replaces the `READ` instruction.

As an aside, this change fixes an issue with `ALTER` in which it did not interpret its instruction index MOD the number of instructions correctly.