Extracting numbers fails in the right conditions
If a value is extracted at the end of the instruction the end index of the extraction is improperly reduced by -1. The reason is that `extraItem` attempts to skip whitespace by preemptively moving the end index after an extraction by searching for the next non-whitespace character. However, if the instruction simply ends in trailing whitespace then the search will return -1 and decrement the end index improperly.