Implement `ASSTRING` and `ASNUMBER`
This implements two commands to cast data into strings and numbers, respectively.

Casting data to a numeric format is straightforward: `UNDEFINED` and lists are always `NAN` and strings which can be parsed as a STOP number are interpreted as that number. Numbers, of course, are numbers.

Casting data to a string is similarly straightforward: data is converted to the form it would take if it were a STOP string with the exception of strings since they are already strings. Converting strings would have erroneously added quotes.

This work also fixes several issues in how instructions were generated and how output data was processed. When casting data to create an instruction the data is always converted to a STOP-parsable format. When preparing data for output, if the root data is `UNDEFINED` then the output is the empty string, otherwise it is a STOP-parsable format.