Unable to recurse into direct references in lists
Similar to the issue in which special direct references deep within lists were not decaying properly, normal direct references were not recursed into. This was also because the case in which the reference was deep within a list was not handled.

The behavior has been updated to search deep within lists for the reference when attempting to recurse and when replacing the direct reference with its decayed value.